feat(theme): fix unit tests
