+++ src/hooks/use-render.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { useEffect } from '@nebula.js/stardust';
-import RootComponent from '../components/root';
+import Root from '../components/Root';
 
 type TODO = any;
 
@@ -15,7 +15,7 @@ type UseRenderProps = {
 const UseRender = ({ core }: UseRenderProps) => {
   useEffect(() => {
     if (core?.element && core.layout) {
-      ReactDOM.render(<RootComponent layout={core.layout} />, core.element);
+      ReactDOM.render(<Root layout={core.layout} />, core.element);
     }
   }, [core]);
 };
